Fixed `Editor` syntax highlighting
The `simple-highlighting` mode wasn't highlighting keywords when they were preceded by other values (like coefficients) because.

This is because it was consuming all letters/numbers and then matching on the current buffer, which meant that keywords only matched if they were bounded by spaces.

Now we instead programmatically generate a `RegExp` object based on the `hintList` observable and pass that to `stream.match` for the highlighting condition.