Dynamically render field type info based on edit
In order to make better use of the space in the config window, the `FieldListItem` now dynamically renders the field type component based on whether or not the bound is being edited.

When editing the bound, the field type info is hidden so that the bound component can take full width of the UI. Otherwise, both are rendered in the field list item.